+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2021 - 2021/01/24
+- Updated requirements
+- Fallback to requirements.txt when poetry throws an error and direct the user
+  to troubleshoot
+
 ## 2020.0.4 - 2020/10/14
 - Improved support for GPL fans out there by detecting variants in a more
   granular manner. Fewer false -ves for said GPL variants. E.g. a dependency with